@@ -1,51 +1,60 @@
 {
   "name": "wmail",
-  "version": "1.3.1",
-  "prerelease": false,
-  "description": "The missing desktop client for Gmail & Google Inbox",
+  "version": "2.3.1",
+  "prerelease": true,
+  "description": "The missing desktop client for Gmail and Google Inbox",
   "scripts": {
     "prestart": "webpack",
     "start": "electron bin/app/index.js",
     "test": "standard",
-    "package:darwin": "node packager.js darwin",
-    "package:linux": "node packager.js linux",
-    "package:win32": "node packager.js win32",
-    "package": "node packager.js darwin && node packager.js linux && node packager.js win32",
-    "postinstall": "electron-rebuild",
-    "install-all": "echo ':>wmail'; npm install; cd src/app; echo ':>wmail-app'; npm install; cd ../../src/scenes/mailboxes; echo ':>wmail-scenes-mailboxes'; npm install",
-    "outdated-all": "echo ':>wmail'; npm outdated; cd src/app; echo ':>wmail-app'; npm outdated; cd ../../src/scenes/mailboxes; echo ':>wmail-scenes-mailboxes'; npm outdated"
+    "install-all": "echo ':wmail' && npm install && cd src/app && echo ':wmail-app' && npm install && cd ../../src/scenes/mailboxes && echo ':wmail-scenes-mailboxes' && npm install",
+    "outdated-all": "echo ':wmail' && npm outdated && cd src/app && echo ':wmail-app' && npm outdated && cd ../../src/scenes/mailboxes && echo ':wmail-scenes-mailboxes' && npm outdated",
+    "dev:platform": "webpack --task=platform && electron bin/app/index.js",
+    "dev:app": "webpack --task=app && electron bin/app/index.js",
+    "dev:mailboxes": "webpack --task=mailboxes && electron bin/app/index.js",
+    "dev:assets": "webpack --task=assets && electron bin/app/index.js",
+    "dev:run": "electron bin/app/index.js"
   },
   "keywords": [],
-  "author": "Thomas Beverley",
-  "license": "BSD-2-Clause",
+  "author": {
+    "name": "Thomas Beverley",
+    "email": "tom.beverley.wmail@gmail.com",
+    "url": "https://github.com/Thomas101/"
+  },
+  "homepage": "https://thomas101.github.io/wmail/",
+  "license": "MPL-2.0",
   "repository": "https://github.com/Thomas101/wmail",
   "main": "bin/app/index.js",
   "dependencies": {
-    "babel": "6.5.2",
-    "babel-core": "6.9.1",
-    "babel-loader": "6.2.4",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-react": "6.5.0",
-    "babel-preset-stage-0": "6.5.0",
-    "clean-webpack-plugin": "0.1.9",
-    "copy-webpack-plugin": "3.0.1",
-    "css-loader": "0.23.1",
-    "electron-packager": "7.0.2",
-    "electron-prebuilt": "1.2.0",
-    "electron-rebuild": "1.1.5",
-    "electron-winstaller": "2.3.0",
+    "babel": "6.23.0",
+    "babel-core": "6.23.1",
+    "babel-loader": "6.3.2",
+    "babel-preset-es2015": "6.22.0",
+    "babel-preset-react": "6.23.0",
+    "babel-preset-stage-0": "6.22.0",
+    "clean-webpack-plugin": "0.1.15",
+    "copy-webpack-plugin": "4.0.1",
+    "css-loader": "0.26.1",
+    "electron": "1.6.1",
     "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.8.5",
+    "file-loader": "0.10.0",
+    "json-loader": "0.5.4",
     "jsx-loader": "0.13.2",
-    "less": "2.7.1",
+    "less": "2.7.2",
     "less-loader": "2.2.3",
-    "nlf": "1.4.0",
     "style-loader": "0.13.1",
+    "uglify-js": "mishoo/UglifyJS2#3ee46e91e802fb8bf20656bce115375c5f624052",
     "url-loader": "0.5.7",
-    "webpack": "1.13.1",
+    "uuid": "3.0.1",
+    "webpack": "1.14.0",
     "webpack-target-electron-renderer": "0.4.0"
   },
   "devDependencies": {
-    "standard": "^7.1.1"
+    "standard": "8.6.0"
+  },
+  "standard": {
+    "ignore": [
+      "src/app/lib/"
+    ]
   }
 }
