@@ -19,7 +19,8 @@ module.exports = {
     }),
     new CopyWebpackPlugin([
       { from: path.join(__dirname, 'fonts'), to: 'fonts', force: true },
-      { from: path.join(__dirname, 'icons'), to: 'icons', force: true }
+      { from: path.join(__dirname, 'icons'), to: 'icons', force: true },
+      { from: path.join(__dirname, 'images'), to: 'images', force: true }
     ], {
       ignore: [ '.DS_Store' ]
     })
