@@ -36,41 +36,48 @@ class AppPrimaryMenu {
           { type: 'separator' },
           process.platform === 'darwin' ? { label: 'Services', role: 'services', submenu: [] } : undefined,
           process.platform === 'darwin' ? { type: 'separator' } : undefined,
-          { label: 'Show Window', accelerator: 'Command+N', click: this._selectors.showWindow },
-          { label: 'Hide Window', accelerator: 'Command+W', click: this._selectors.closeWindow },
+          { label: 'Show Window', accelerator: 'CmdOrCtrl+N', click: this._selectors.showWindow },
+          { label: 'Hide Window', accelerator: 'CmdOrCtrl+W', click: this._selectors.closeWindow },
           { label: 'Hide', accelerator: 'CmdOrCtrl+H', role: 'hide' },
-          { label: 'Hide Others', accelerator: 'Alt+CmdOrCtrl+H', role: 'hideothers' },
+          { label: 'Hide Others', accelerator: process.platform === 'darwin' ? 'Command+Alt+H' : 'Ctrl+Shift+H', role: 'hideothers' },
           { label: 'Show All', role: 'unhide' },
           { type: 'separator' },
-          { label: 'Quit', accelerator: 'Command+Q', click: this._selectors.fullQuit }
+          { label: 'Quit', accelerator: 'CmdOrCtrl+Q', click: this._selectors.fullQuit }
         ].filter((item) => item !== undefined)
       },
       {
         label: 'Edit',
         submenu: [
-          { label: 'Undo', accelerator: 'CmdOrCtrl+Z', selector: 'undo:' },
-          { label: 'Redo', accelerator: 'Shift+CmdOrCtrl+Z', selector: 'redo:' },
+          { label: 'Undo', accelerator: 'CmdOrCtrl+Z', role: 'undo' },
+          { label: 'Redo', accelerator: 'Shift+CmdOrCtrl+Z', role: 'redo' },
           { type: 'separator' },
-          { label: 'Cut', accelerator: 'CmdOrCtrl+X', selector: 'cut:' },
-          { label: 'Copy', accelerator: 'CmdOrCtrl+C', selector: 'copy:' },
-          { label: 'Paste', accelerator: 'CmdOrCtrl+V', selector: 'paste:' },
-          { label: 'Paste and match style', accelerator: 'Command+Shift+V', selector: 'pasteAndMatchStyle:' },
-          { label: 'Select All', accelerator: 'CmdOrCtrl+A', selector: 'selectAll:' }
+          { label: 'Cut', accelerator: 'CmdOrCtrl+X', role: 'cut' },
+          { label: 'Copy', accelerator: 'CmdOrCtrl+C', role: 'copy' },
+          { label: 'Paste', accelerator: 'CmdOrCtrl+V', role: 'paste' },
+          { label: 'Paste and match style', accelerator: 'CmdOrCtrl+Shift+V', role: 'pasteandmatchstyle' },
+          { label: 'Select All', accelerator: 'CmdOrCtrl+A', role: 'selectall' },
+          { type: 'separator' },
+          { label: 'Find', accelerator: 'CmdOrCtrl+F', click: this._selectors.find },
+          { label: 'Find Next', accelerator: 'CmdOrCtrl+G', click: this._selectors.findNext }
         ]
       },
       {
         label: 'View',
         submenu: [
-          { label: 'Toggle Full Screen', accelerator: 'Ctrl+Command+F', click: this._selectors.fullscreenToggle },
-          { label: 'Toggle Sidebar', accelerator: 'Ctrl+Command+S', click: this._selectors.sidebarToggle },
+          { label: 'Toggle Full Screen', accelerator: process.platform === 'darwin' ? 'Ctrl+Command+F' : 'F11', click: this._selectors.fullscreenToggle },
+          { label: 'Toggle Sidebar', accelerator: (process.platform === 'darwin' ? 'Ctrl+Command+S' : 'Ctrl+Shift+S'), click: this._selectors.sidebarToggle },
           process.platform === 'darwin' ? undefined : { label: 'Toggle Menu', accelerator: 'CmdOrCtrl+\\', click: this._selectors.menuToggle },
+          { type: 'separator' },
+          { label: 'Navigate Back', accelerator: 'CmdOrCtrl+[', click: this._selectors.mailboxNavBack },
+          { label: 'Navigate Back', accelerator: 'CmdOrCtrl+Left', click: this._selectors.mailboxNavBack },
+          { label: 'Navigate Forward', accelerator: 'CmdOrCtrl+]', click: this._selectors.mailboxNavForward },
           { type: 'separator' },
           { label: 'Zoom Mailbox In', accelerator: 'CmdOrCtrl+Plus', click: this._selectors.zoomIn },
           { label: 'Zoom Mailbox Out', accelerator: 'CmdOrCtrl+-', click: this._selectors.zoomOut },
-          { label: 'Reset Mailbox Zoom', click: this._selectors.zoomReset },
+          { label: 'Reset Mailbox Zoom', accelerator: 'CmdOrCtrl+0', click: this._selectors.zoomReset },
           { type: 'separator' },
           { label: 'Reload', accelerator: 'CmdOrCtrl+R', click: this._selectors.reload },
-          { label: 'Developer Tools', accelerator: 'Alt+CmdOrCtrl+J', click: this._selectors.devTools }
+          { label: 'Developer Tools', accelerator: process.platform === 'darwin' ? 'Cmd+Alt+J' : 'Ctrl+Shift+J', click: this._selectors.devTools }
         ].filter((item) => item !== undefined)
       },
       {
@@ -80,16 +87,23 @@ class AppPrimaryMenu {
           { label: 'Minimize', accelerator: 'CmdOrCtrl+M', role: 'minimize' },
           { label: 'Cycle Windows', accelerator: 'CmdOrCtrl+`', click: this._selectors.cycleWindows }
         ]
-        .concat(mailboxes.length ? [{ type: 'separator' }] : [])
-        .concat(mailboxes.map((mailbox, index) => {
-          return { label: mailbox.email || 'Untitled', accelerator: 'CmdOrCtrl+' + (index + 1), click: () => { this._selectors.mailbox(mailbox.id) } }
+        .concat(mailboxes.length <= 1 ? [] : [
+          { type: 'separator' },
+          { label: 'Previous Mailbox', accelerator: 'CmdOrCtrl+<', click: this._selectors.prevMailbox },
+          { label: 'Next Mailbox', accelerator: 'CmdOrCtrl+>', click: this._selectors.nextMailbox },
+          { type: 'separator' }
+        ])
+        .concat(mailboxes.length <= 1 ? [] : mailboxes.map((mailbox, index) => {
+          return { label: mailbox.email || 'Untitled', accelerator: 'CmdOrCtrl+' + (index + 1), click: () => { this._selectors.changeMailbox(mailbox.id) } }
         }))
       },
       {
         label: 'Help',
         role: 'help',
         submenu: [
-          { label: 'Project Homepage', click: this._selectors.learnMore },
+          { label: 'WMail Website', click: this._selectors.learnMore },
+          { label: 'Privacy', click: this._selectors.privacy },
+          { label: 'WMail on GitHub', click: this._selectors.learnMoreGithub },
           { label: 'Report a Bug', click: this._selectors.bugReport }
         ]
       }
@@ -124,6 +138,18 @@ class AppPrimaryMenu {
         this.updateApplicationMenu(nextMailboxes)
       }
     }
+  }
+
+  /* ****************************************************************************/
+  // Click handlers
+  /* ****************************************************************************/
+
+  changeToPrevMailbox () {
+
+  }
+
+  changeToNextMailbox () {
+
   }
 }
 
