@@ -1,9 +1,12 @@
+const {ipcMain} = require('electron')
 const AppDirectory = require('appdirectory')
 const pkg = require('../../package.json')
 const mkdirp = require('mkdirp')
-const Storage = require('dom-storage')
 const path = require('path')
 const Minivents = require('minivents')
+const fs = require('fs')
+const writeFileAtomic = require('write-file-atomic')
+const { DB_WRITE_DELAY_MS } = require('../../shared/constants')
 
 // Setup
 const appDirectory = new AppDirectory(pkg.name)
@@ -17,8 +20,62 @@ class StorageBucket {
   /* ****************************************************************************/
 
   constructor (bucketName) {
-    this.__storage__ = new Storage(path.join(dbPath, bucketName + '_db.json'))
+    this.__path__ = path.join(dbPath, bucketName + '_db.json')
+    this.__writeHold__ = null
+    this.__writeLock__ = false
+    this.__data__ = undefined
+    this.__ipcReplyChannel__ = `storageBucket:${bucketName}:reply`
+
+    this._loadFromDiskSync()
+
+    ipcMain.on(`storageBucket:${bucketName}:setItem`, this._handleIPCSetItem.bind(this))
+    ipcMain.on(`storageBucket:${bucketName}:removeItem`, this._handleIPCRemoveItem.bind(this))
+    ipcMain.on(`storageBucket:${bucketName}:getItem`, this._handleIPCGetItem.bind(this))
+    ipcMain.on(`storageBucket:${bucketName}:allKeys`, this._handleIPCAllKeys.bind(this))
+    ipcMain.on(`storageBucket:${bucketName}:allItems`, this._handleIPCAllItems.bind(this))
+
     Minivents(this)
+  }
+
+  checkAwake () { return true }
+
+  /* ****************************************************************************/
+  // Persistence
+  /* ****************************************************************************/
+
+  /**
+  * Loads the database from disk
+  */
+  _loadFromDiskSync () {
+    let data = '{}'
+    try {
+      data = fs.readFileSync(this.__path__, 'utf8')
+    } catch (ex) { }
+
+    try {
+      this.__data__ = JSON.parse(data)
+    } catch (ex) {
+      this.__data__ = {}
+    }
+  }
+
+  /**
+  * Writes the current data to disk
+  */
+  _writeToDisk () {
+    clearTimeout(this.__writeHold__)
+    this.__writeHold__ = setTimeout(() => {
+      if (this.__writeLock__) {
+        // Requeue in DB_WRITE_DELAY_MS
+        this._writeToDisk()
+        return
+      } else {
+        this.__writeLock__ = true
+        writeFileAtomic(this.__path__, JSON.stringify(this.__data__), () => {
+          this.__writeLock__ = false
+        })
+      }
+    }, DB_WRITE_DELAY_MS)
   }
 
   /* ****************************************************************************/
@@ -28,11 +85,11 @@ class StorageBucket {
   /**
   * @param k: the key of the item
   * @param d=undefined: the default value if not exists
-  * @return the json item or d
+  * @return the string item or d
   */
   getItem (k, d) {
-    const json = this.__storage__.getItem(k)
-    return json ? JSON.parse(json) : d
+    const json = this.__data__[k]
+    return json || d
   }
 
   /**
@@ -40,20 +97,20 @@ class StorageBucket {
   * @param d=undefined: the default value if not exists
   * @return the string item or d
   */
-  getString (k, d) {
-    const json = this.__storage__.getItem(k)
-    return json || d
+  getJSONItem (k, d) {
+    const item = this.getItem(k)
+    try {
+      return item ? JSON.parse(item) : d
+    } catch (ex) {
+      return {}
+    }
   }
 
   /**
   * @return a list of all keys
   */
   allKeys () {
-    const keys = []
-    for (let i = 0; i < this.__storage__.length; i++) {
-      keys.push(this.__storage__.key(i))
-    }
-    return keys
+    return Object.keys(this.__data__)
   }
 
   /**
@@ -67,17 +124,17 @@ class StorageBucket {
   }
 
   /**
-  * @return all the items in an obj
+  * @return all the items in an obj json parsed
   */
-  allStrings () {
+  allJSONItems () {
     return this.allKeys().reduce((acc, key) => {
-      acc[key] = this.getString(key)
+      acc[key] = this.getJSONItem(key)
       return acc
     }, {})
   }
 
   /* ****************************************************************************/
-  // Setters
+  // Modifiers
   /* ****************************************************************************/
 
   /**
@@ -85,36 +142,94 @@ class StorageBucket {
   * @param v: the value to set
   * @return v
   */
-  setItem (k, v) {
-    this.__storage__.setItem(k, JSON.stringify(v))
+  _setItem (k, v) {
+    this.__data__[k] = '' + v
+    this._writeToDisk()
     this.emit('changed', { type: 'setItem', key: k })
     this.emit('changed:' + k, { })
     return v
   }
 
   /**
-  * @param k: the key to set
-  * @param s: the value to set
-  * @return s
+  * @param k: the key to remove
   */
-  setString (k, s) {
-    this.__storage__.setItem(k, s)
-    this.emit('changed', { type: 'setString', key: k })
+  _removeItem (k) {
+    delete this.__data__[k]
+    this._writeToDisk()
+    this.emit('changed', { type: 'removeItem', key: k })
     this.emit('changed:' + k, { })
-    return s
   }
 
   /* ****************************************************************************/
-  // Removers
+  // IPC Access
   /* ****************************************************************************/
 
   /**
-  * @param k: the key to remove
+  * Responds to an ipc message
+  * @param evt: the original event that fired
+  * @param response: teh response to send
+  * @param sendSync: set to true to respond synchronously
   */
-  removeItem (k) {
-    this.__storage__.removeItem(k)
-    this.emit('changed', { type: 'removeItem', key: k })
-    this.emit('changed:' + k, { })
+  _sendIPCResponse (evt, response, sendSync = false) {
+    if (sendSync) {
+      evt.returnValue = response
+    } else {
+      evt.sender.send(this.__ipcReplyChannel__, response)
+    }
+  }
+
+  /**
+  * Sets an item over IPC
+  * @param evt: the fired event
+  * @param body: request body
+  */
+  _handleIPCSetItem (evt, body) {
+    this._setItem(body.key, body.value)
+    this._sendIPCResponse(evt, { id: body.id, response: null }, body.sync)
+  }
+
+  /**
+  * Removes an item over IPC
+  * @param evt: the fired event
+  * @param body: request body
+  */
+  _handleIPCRemoveItem (evt, body) {
+    this._removeItem(body.key)
+    this._sendIPCResponse(evt, { id: body.id, response: null }, body.sync)
+  }
+
+  /**
+  * Gets an item over IPC
+  * @param evt: the fired event
+  * @param body: request body
+  */
+  _handleIPCGetItem (evt, body) {
+    this._sendIPCResponse(evt, {
+      id: body.id,
+      response: this.getItem(body.key)
+    }, body.sync)
+  }
+
+  /**
+  * Gets the keys over IPC
+  * @param body: request body
+  */
+  _handleIPCAllKeys (evt, body) {
+    this._sendIPCResponse(evt, {
+      id: body.id,
+      response: this.allKeys()
+    }, body.sync)
+  }
+
+  /**
+  * Gets all the items over IPC
+  * @param body: request body
+  */
+  _handleIPCAllItems (evt, body) {
+    this._sendIPCResponse(evt, {
+      id: body.id,
+      response: this.allItems()
+    }, body.sync)
   }
 }
 
