@@ -4,6 +4,10 @@ const Mailbox = require('../../shared/Models/Mailbox/Mailbox')
 const { MAILBOX_INDEX_KEY } = require('../../shared/constants')
 
 class MailboxStore {
+  /* ****************************************************************************/
+  // Lifecycle
+  /* ****************************************************************************/
+
   constructor () {
     Minivents(this)
 
@@ -11,7 +15,7 @@ class MailboxStore {
     this.index = []
     this.mailboxes = new Map()
 
-    const allRawItems = persistence.allItems()
+    const allRawItems = persistence.allJSONItems()
     Object.keys(allRawItems).forEach((id) => {
       if (id === MAILBOX_INDEX_KEY) {
         this.index = allRawItems[id]
@@ -23,10 +27,10 @@ class MailboxStore {
     // Listen for changes
     persistence.on('changed', (evt) => {
       if (evt.key === MAILBOX_INDEX_KEY) {
-        this.index = persistence.getItem(MAILBOX_INDEX_KEY)
+        this.index = persistence.getJSONItem(MAILBOX_INDEX_KEY)
       } else {
         if (evt.type === 'setItem') {
-          this.mailboxes.set(evt.key, new Mailbox(evt.key, persistence.getItem(evt.key)))
+          this.mailboxes.set(evt.key, new Mailbox(evt.key, persistence.getJSONItem(evt.key)))
         }
         if (evt.type === 'removeItem') {
           this.mailboxes.delete(evt.key)
@@ -36,10 +40,25 @@ class MailboxStore {
     })
   }
 
+  /* ****************************************************************************/
+  // Getters
+  /* ****************************************************************************/
+
+  /**
+  * @return the mailboxes in an ordered list
+  */
   orderedMailboxes () {
     return this.index
       .map(id => this.mailboxes.get(id))
       .filter((mailbox) => !!mailbox)
+  }
+
+  /**
+  * @param id: the id of the mailbox
+  * @return the mailbox record
+  */
+  getMailbox (id) {
+    return this.mailboxes.get(id)
   }
 }
 
