@@ -7,8 +7,14 @@ class ContentWindow extends WMailWindow {
   // Creation
   /* ****************************************************************************/
 
-  defaultWindowPreferences (partition) {
-    return Object.assign(super.defaultWindowPreferences(), {
+  /**
+  * The default window preferences
+  * @param partition: the partition to set the window to
+  * @param extraPreferences = undefined: extra preferences to merge into the prefs
+  * @return the settings
+  */
+  defaultWindowPreferences (partition, extraPreferences = undefined) {
+    return Object.assign(super.defaultWindowPreferences(extraPreferences), {
       minWidth: 400,
       minHeight: 400,
       webPreferences: {
@@ -18,10 +24,20 @@ class ContentWindow extends WMailWindow {
     })
   }
 
-  start (url, partition) {
-    this.createWindow(this.defaultWindowPreferences(partition), url)
+  /**
+  * Starts the window
+  * @param url: the start url
+  * @param partition: the window partition
+  * @param windowPreferences=undefined: additional window preferences to supply
+  */
+  start (url, partition, windowPreferences = undefined) {
+    this.createWindow(this.defaultWindowPreferences(partition, windowPreferences), url)
   }
 
+  /**
+  * Creates and launches the window
+  * @arguments: passed through to super()
+  */
   createWindow () {
     super.createWindow.apply(this, Array.from(arguments))
     this.window.webContents.on('new-window', (evt, url) => {
