@@ -1,13 +1,23 @@
+import './ReactComponents.less'
 const React = require('react')
 const ReactDOM = require('react-dom')
 const App = require('./ui/App')
 const mailboxActions = require('./stores/mailbox/mailboxActions')
 const settingsActions = require('./stores/settings/settingsActions')
-const ipc = window.nativeRequire('electron').ipcRenderer
+const composeActions = require('./stores/compose/composeActions')
+const mailboxWizardActions = require('./stores/mailboxWizard/mailboxWizardActions')
+const { ipcRenderer } = window.nativeRequire('electron')
+
+// See if we're offline and run a re-direct
+if (window.navigator.onLine === false) {
+  window.location.href = 'offline.html'
+}
 
 // Load what we have in the db
 mailboxActions.load()
+mailboxWizardActions.load()
 settingsActions.load()
+composeActions.load()
 
 // Remove loading
 ;(() => {
@@ -17,6 +27,7 @@ settingsActions.load()
 
 // Render and prepare for unrender
 ReactDOM.render(<App />, document.getElementById('app'))
-ipc.on('prepare-reload', function () {
+ipcRenderer.on('prepare-reload', function () {
   ReactDOM.unmountComponentAtNode(document.getElementById('app'))
 })
+ipcRenderer.send('mailboxes-js-loaded', {})
