@@ -1,12 +1,21 @@
 const React = require('react')
 const { Toggle, TextField, Paper } = require('material-ui')
-const { Row, Col } = require('../../Components/Flexbox')
+const { Container, Row, Col } = require('../../Components/Grid')
 const flux = {
   settings: require('../../stores/settings')
 }
+const styles = require('./settingStyles')
+const shallowCompare = require('react-addons-shallow-compare')
 
 module.exports = React.createClass({
+  /* **************************************************************************/
+  // Class
+  /* **************************************************************************/
+
   displayName: 'AdvancedSettings',
+  propTypes: {
+    showRestart: React.PropTypes.func.isRequired
+  },
 
   /* **************************************************************************/
   // Lifecycle
@@ -32,7 +41,8 @@ module.exports = React.createClass({
     return {
       proxyEnabled: store.proxy.enabled,
       proxyHost: store.proxy.host || '',
-      proxyPort: store.proxy.port || ''
+      proxyPort: store.proxy.port || '',
+      app: store.app
     }
   },
 
@@ -66,39 +76,80 @@ module.exports = React.createClass({
   // Rendering
   /* **************************************************************************/
 
-  /**
-  * Renders the app
-  */
+  shouldComponentUpdate (nextProps, nextState) {
+    return shallowCompare(this, nextProps, nextState)
+  },
+
   render () {
+    const { proxyEnabled, proxyPort, proxyHost, app } = this.state
+    const { showRestart, ...passProps } = this.props
+
     return (
-      <div {...this.props}>
-        <Paper zDepth={1} style={{ padding: 15 }}>
+      <div {...passProps}>
+        <Paper zDepth={1} style={styles.paper}>
+          <h1 style={styles.subheading}>Proxy Server</h1>
           <Toggle
             name='proxyEnabled'
-            defaultToggled={this.state.proxyEnabled}
+            defaultToggled={proxyEnabled}
             label='Enable Proxy Server'
             onToggle={this.handleProxyToggle} />
           <small>You also need to set the proxy settings on your OS to ensure all requests use the server</small>
-          <Row>
-            <Col xs={6}>
-              <TextField
-                ref='proxy_host'
-                hintText='http://192.168.1.1'
-                floatingLabelText='Proxy Server Host'
-                defaultValue={this.state.proxyHost}
-                onChange={this.handleProxyValueChanged}
-                disabled={!this.state.proxyEnabled} />
-            </Col>
-            <Col xs={6}>
-              <TextField
-                ref='proxy_port'
-                hintText='8080'
-                floatingLabelText='Proxy Server Port'
-                defaultValue={this.state.proxyPort}
-                onChange={this.handleProxyValueChanged}
-                disabled={!this.state.proxyEnabled} />
-            </Col>
-          </Row>
+          <Container fluid>
+            <Row>
+              <Col xs={6}>
+                <TextField
+                  ref='proxy_host'
+                  hintText='http://192.168.1.1'
+                  floatingLabelText='Proxy Server Host'
+                  defaultValue={proxyHost}
+                  onChange={this.handleProxyValueChanged}
+                  disabled={!proxyEnabled} />
+              </Col>
+              <Col xs={6}>
+                <TextField
+                  ref='proxy_port'
+                  hintText='8080'
+                  floatingLabelText='Proxy Server Port'
+                  defaultValue={proxyPort}
+                  onChange={this.handleProxyValueChanged}
+                  disabled={!proxyEnabled} />
+              </Col>
+            </Row>
+          </Container>
+        </Paper>
+        <Paper zDepth={1} style={styles.paper}>
+          <Toggle
+            toggled={app.ignoreGPUBlacklist}
+            label='Ignore GPU Blacklist (Requires Restart)'
+            labelPosition='right'
+            onToggle={(evt, toggled) => {
+              showRestart()
+              flux.settings.A.ignoreGPUBlacklist(toggled)
+            }} />
+          <Toggle
+            toggled={app.enableUseZoomForDSF}
+            label='Use Zoom For DSF (Requires Restart)'
+            labelPosition='right'
+            onToggle={(evt, toggled) => {
+              showRestart()
+              flux.settings.A.enableUseZoomForDSF(toggled)
+            }} />
+          <Toggle
+            toggled={app.disableSmoothScrolling}
+            label='Disable Smooth Scrolling (Requires Restart)'
+            labelPosition='right'
+            onToggle={(evt, toggled) => {
+              showRestart()
+              flux.settings.A.disableSmoothScrolling(toggled)
+            }} />
+          <Toggle
+            toggled={app.checkForUpdates}
+            label='Check for updates'
+            labelPosition='right'
+            onToggle={(evt, toggled) => {
+              showRestart()
+              flux.settings.A.checkForUpdates(toggled)
+            }} />
         </Paper>
       </div>
     )
