@@ -7,9 +7,34 @@ const GeneralSettings = require('./GeneralSettings')
 const AccountSettings = require('./AccountSettings')
 const AdvancedSettings = require('./AdvancedSettings')
 const Colors = require('material-ui/styles/colors')
+const styles = require('./settingStyles')
+const { ipcRenderer } = window.nativeRequire('electron')
 
 module.exports = React.createClass({
+  /* **************************************************************************/
+  // Class
+  /* **************************************************************************/
+
   displayName: 'SettingsDialog',
+  propTypes: {
+    open: React.PropTypes.bool.isRequired,
+    onRequestClose: React.PropTypes.func.isRequired,
+    initialRoute: React.PropTypes.object
+  },
+
+  /* **************************************************************************/
+  // Component Lifecycle
+  /* **************************************************************************/
+
+  componentWillReceiveProps (nextProps) {
+    if (this.props.open !== nextProps.open) {
+      const updates = { showRestart: false }
+      if (nextProps.open) {
+        updates.currentTab = (nextProps.initialRoute || {}).tab || 'general'
+      }
+      this.setState(updates)
+    }
+  },
 
   /* **************************************************************************/
   // Data lifecycle
@@ -17,15 +42,9 @@ module.exports = React.createClass({
 
   getInitialState () {
     return {
-      currentTab: 'general'
+      currentTab: (this.props.initialRoute || {}).tab || 'general',
+      showRestart: false
     }
-  },
-
-  shouldComponentUpdate (nextProps, nextState) {
-    if (this.state.currentTab !== nextState.currentTab) { return true }
-    if (nextProps.open !== this.props.open) { return true }
-
-    return false
   },
 
   /* **************************************************************************/
@@ -42,48 +61,57 @@ module.exports = React.createClass({
   },
 
   /**
-  * Closes the modal
+  * Shows the option to restart
   */
-  handleClose () {
-    this.props.onRequestClose()
+  handleShowRestart () {
+    this.setState({ showRestart: true })
   },
 
   /* **************************************************************************/
   // Rendering
   /* **************************************************************************/
 
-  /**
-  * Renders the app
-  */
+  shouldComponentUpdate (nextProps, nextState) {
+    if (this.state.currentTab !== nextState.currentTab) { return true }
+    if (this.state.showRestart !== nextState.showRestart) { return true }
+    if (nextProps.open !== this.props.open) { return true }
+
+    return false
+  },
+
   render () {
-    const buttons = (
+    const { showRestart, currentTab } = this.state
+    const { onRequestClose, initialRoute, open } = this.props
+
+    const buttons = showRestart ? (
       <div style={{ textAlign: 'right' }}>
-        <RaisedButton label='Close' primary onClick={this.handleClose} />
+        <RaisedButton label='Close' style={{ marginRight: 16 }} onClick={onRequestClose} />
+        <RaisedButton label='Restart' primary onClick={() => ipcRenderer.send('relaunch-app', { })} />
+      </div>
+    ) : (
+      <div style={{ textAlign: 'right' }}>
+        <RaisedButton label='Close' primary onClick={onRequestClose} />
       </div>
     )
 
-    const tabInfo = [
-      { label: 'General', value: 'general' },
-      { label: 'Accounts', value: 'accounts' },
-      { label: 'Advanced', value: 'advanced' }
+    const tabHeadings = [
+      ['General', 'general'],
+      ['Accounts', 'accounts'],
+      ['Advanced', 'advanced']
     ]
+
     const heading = (
-      <div style={{ display: 'flex', flexDirection: 'row', alignContent: 'stretch' }}>
-        {tabInfo.map(({label, value}) => {
+      <div style={styles.tabToggles}>
+        {tabHeadings.map(([label, value]) => {
           return (
             <FlatButton
               key={value}
               label={label}
-              style={{
-                height: 50,
-                borderRadius: 0,
-                flex: 1,
-                borderBottomWidth: 2,
-                borderBottomStyle: 'solid',
-                borderBottomColor: this.state.currentTab === value ? Colors.redA200 : 'transparent'
-              }}
+              style={Object.assign({}, styles.tabToggle, {
+                borderBottomColor: currentTab === value ? Colors.redA200 : 'transparent'
+              })}
               labelStyle={{
-                color: this.state.currentTab === value ? Colors.white : Colors.lightBlue100
+                color: currentTab === value ? Colors.white : Colors.lightBlue100
               }}
               backgroundColor={Colors.lightBlue600}
               hoverColor={Colors.lightBlue600}
@@ -97,27 +125,36 @@ module.exports = React.createClass({
     return (
       <Dialog
         modal={false}
+        contentStyle={styles.dialog}
         title={heading}
         actions={buttons}
-        open={this.props.open}
+        open={open}
         bodyStyle={{ padding: 0 }}
         titleStyle={{ padding: 0 }}
         autoScrollBodyContent
-        onRequestClose={this.props.onRequestClose}>
+        onRequestClose={onRequestClose}>
         <Tabs
           inkBarStyle={{ display: 'none' }}
           tabItemContainerStyle={{ display: 'none' }}
-          value={this.state.currentTab}
+          value={currentTab}
           onChange={this.handleTabChange}
           contentContainerStyle={{ padding: 24 }}>
           <Tab label='General' value='general'>
-            <GeneralSettings />
+            {currentTab !== 'general' ? undefined : (
+              <GeneralSettings showRestart={this.handleShowRestart} />
+            )}
           </Tab>
           <Tab label='Accounts' value='accounts'>
-            <AccountSettings />
+            {currentTab !== 'accounts' ? undefined : (
+              <AccountSettings
+                showRestart={this.handleShowRestart}
+                initialMailboxId={(initialRoute || {}).mailboxId} />
+            )}
           </Tab>
           <Tab label='Advanced' value='advanced'>
-            <AdvancedSettings />
+            {currentTab !== 'advanced' ? undefined : (
+              <AdvancedSettings showRestart={this.handleShowRestart} />
+            )}
           </Tab>
         </Tabs>
       </Dialog>
