@@ -1,11 +1,28 @@
 const Model = require('../Model')
-const { GMAIL_NOTIFICATION_MAX_MESSAGE_AGE_MS } = require('../../constants')
+const SERVICES = require('./MailboxServices')
+const TYPES = require('./MailboxTypes')
 
 const UNREAD_MODES = {
   INBOX: 'inbox',
   INBOX_UNREAD: 'inbox_unread',
-  PRIMARY_INBOX_UNREAD: 'primary_inbox_unread'
+  PRIMARY_INBOX_UNREAD: 'primary_inbox_unread',
+  INBOX_UNREAD_IMPORTANT: 'inbox_unread_important',
+  GINBOX_DEFAULT: 'ginbox_default'
 }
+
+const SERVICE_URLS = { }
+SERVICE_URLS[SERVICES.STORAGE] = 'https://drive.google.com'
+SERVICE_URLS[SERVICES.CONTACTS] = 'https://contacts.google.com'
+SERVICE_URLS[SERVICES.NOTES] = 'https://keep.google.com'
+SERVICE_URLS[SERVICES.CALENDAR] = 'https://calendar.google.com'
+SERVICE_URLS[SERVICES.COMMUNICATION] = 'https://hangouts.google.com'
+
+const SERVICE_NAMES = { }
+SERVICE_NAMES[SERVICES.STORAGE] = 'Drive'
+SERVICE_NAMES[SERVICES.CONTACTS] = 'Contacts'
+SERVICE_NAMES[SERVICES.NOTES] = 'Notes'
+SERVICE_NAMES[SERVICES.CALENDAR] = 'Calendar'
+SERVICE_NAMES[SERVICES.COMMUNICATION] = 'Hangouts'
 
 class Google extends Model {
 
@@ -14,18 +31,30 @@ class Google extends Model {
   /* **************************************************************************/
 
   static get UNREAD_MODES () { return UNREAD_MODES }
+  static get SUPPORTED_SERVICES () { return Object.keys(SERVICES).map((k) => SERVICES[k]) }
+  static get DEFAULT_SERVICES () { return [SERVICES.CALENDAR, SERVICES.STORAGE, SERVICES.NOTES] }
+  static get SERVICE_URLS () { return SERVICE_URLS }
+  static get SERVICE_NAMES () { return SERVICE_NAMES }
 
   /* **************************************************************************/
   // Lifecycle
   /* **************************************************************************/
 
-  constructor (auth, config, unread) {
+  constructor (type, auth, config, labelInfo, unreadMessageInfo) {
     super({
       auth: auth || {},
       config: config || {},
-      unread: unread || {}
+      labelInfo: labelInfo,
+      unreadMessages: unreadMessageInfo || {}
     })
+    this.__type__ = type
   }
+
+  /* **************************************************************************/
+  // Properties
+  /* **************************************************************************/
+
+  get type () { return this.__type__ }
 
   /* **************************************************************************/
   // Properties : GoogleAuth
@@ -36,17 +65,32 @@ class Google extends Model {
   get accessToken () { return this.__data__.auth.access_token }
   get refreshToken () { return this.__data__.auth.refresh_token }
   get authExpiryTime () { return (this.__data__.auth.date || 0) + (this.__data__.auth.expires_in || 0) }
+  get authHasGrantError () {
+    return this.__data__.auth.invalidGrant === undefined ? false : this.__data__.auth.invalidGrant
+  }
 
   /* **************************************************************************/
   // Properties : Google Config
   /* **************************************************************************/
 
-  get unreadMode () { return this.__data__.config.unreadMode || UNREAD_MODES.INBOX_UNREAD }
+  get unreadMode () {
+    if (this.__data__.config.unreadMode) {
+      return this.__data__.config.unreadMode
+    } else {
+      if (this.type === TYPES.GMAIL) {
+        return UNREAD_MODES.INBOX_UNREAD
+      } else if (this.type === TYPES.GINBOX) {
+        return UNREAD_MODES.GINBOX_DEFAULT
+      }
+    }
+  }
   get unreadQuery () {
     switch (this.unreadMode) {
       case UNREAD_MODES.INBOX: return 'label:inbox'
       case UNREAD_MODES.INBOX_UNREAD: return 'label:inbox label:unread'
       case UNREAD_MODES.PRIMARY_INBOX_UNREAD: return 'label:inbox label:unread category:primary'
+      case UNREAD_MODES.INBOX_UNREAD_IMPORTANT: return 'label:inbox label:unread label:important'
+      case UNREAD_MODES.GINBOX_DEFAULT: return 'label:inbox label:unread -has:userlabels -category:promotions -category:forums -category:social'
     }
   }
   get unreadLabel () {
@@ -54,45 +98,86 @@ class Google extends Model {
       case UNREAD_MODES.INBOX: return 'INBOX'
       case UNREAD_MODES.INBOX_UNREAD: return 'INBOX'
       case UNREAD_MODES.PRIMARY_INBOX_UNREAD: return 'CATEGORY_PERSONAL' // actually primary
+      case UNREAD_MODES.INBOX_UNREAD_IMPORTANT: return 'IMPORTANT'
+      case UNREAD_MODES.GINBOX_DEFAULT: return 'INBOX'
     }
   }
-  get unreadLabelField () {
+  get unreadCountIncludesReadMessages () {
     switch (this.unreadMode) {
-      case UNREAD_MODES.INBOX: return 'threadsTotal'
-      case UNREAD_MODES.INBOX_UNREAD: return 'threadsUnread'
-      case UNREAD_MODES.PRIMARY_INBOX_UNREAD: return 'threadsUnread'
+      case UNREAD_MODES.INBOX: return true
+      case UNREAD_MODES.INBOX_UNREAD: return false
+      case UNREAD_MODES.PRIMARY_INBOX_UNREAD: return false
+      case UNREAD_MODES.INBOX_UNREAD_IMPORTANT: return false
+      case UNREAD_MODES.GINBOX_DEFAULT: return false
     }
   }
-
-  /* **************************************************************************/
-  // Properties : Google Unread
-  /* **************************************************************************/
-
-  get unreadMessages () {
-    return Object.keys(this.__data__.unread)
-      .reduce((acc, k) => {
-        if (this.__data__.unread[k].unread) {
-          acc[k] = this.__data__.unread[k]
-        }
-        return acc
-      }, {})
-  }
-
-  get unreadUnotifiedMessages () {
-    const unotified = {}
-    const unread = this.unreadMessages
-    const now = new Date().getTime()
-
-    for (var k in unread) {
-      const info = unread[k]
-      if (info.notified === undefined && info.message) {
-        const messageDate = new Date(parseInt(info.message.internalDate, 10)).getTime()
-        if (now - messageDate < GMAIL_NOTIFICATION_MAX_MESSAGE_AGE_MS) {
-          unotified[k] = info
+  get takeLabelCountFromUI () {
+    if (this.canChangeTakeLabelCountFromUI) {
+      if (this.__data__.config.takeLabelCountFromUI === undefined) {
+        return false
+      } else {
+        return this.__data__.config.takeLabelCountFromUI
+      }
+    } else {
+      if (this.type === TYPES.GMAIL) {
+        if (this.unreadMode === UNREAD_MODES.PRIMARY_INBOX_UNREAD || this.unreadMode === UNREAD_MODES.INBOX_UNREAD_IMPORTANT) {
+          return true
         }
       }
     }
-    return unotified
+    return false
+  }
+  get canChangeTakeLabelCountFromUI () {
+    if (this.type === TYPES.GMAIL) {
+      return this.unreadMode === UNREAD_MODES.INBOX || this.unreadMode === UNREAD_MODES.INBOX_UNREAD
+    } else {
+      return false
+    }
+  }
+
+  /* **************************************************************************/
+  // Properties : Label info
+  /* **************************************************************************/
+
+  get labelInfo () { return this.__data__.labelInfo || {} }
+  get messagesTotal () { return this.labelInfo.messagesTotal || 0 }
+  get messagesUnread () { return this.labelInfo.messagesUnread || 0 }
+  get threadsTotal () { return this.labelInfo.threadsTotal || 0 }
+  get threadsUnread () { return this.labelInfo.threadsUnread || 0 }
+  get unreadCount () {
+    if (this.unreadMode === UNREAD_MODES.GINBOX_DEFAULT) {
+      return this.__data__.unreadMessages.resultSizeEstimate || 0
+    } else {
+      return this.unreadCountIncludesReadMessages ? this.threadsTotal : this.threadsUnread
+    }
+  }
+
+  /* **************************************************************************/
+  // Properties : Unread Messages
+  /* **************************************************************************/
+
+  get latestUnreadThreads () {
+    return this.__data__.unreadMessages.latestUnreadThreads || []
+  }
+
+  get latestUnreadMessages () {
+    return this.latestUnreadThreads.map((thread) => {
+      const messages = thread.messages || []
+      for (var i = messages.length - 1; i >= 0; i--) {
+        const message = messages[i]
+        const wasSent = (message.labelIds || []).findIndex((label) => label === 'SENT') !== -1
+        if (!wasSent) { return message }
+      }
+      return undefined
+    }).filter((m) => !!m)
+  }
+  get unnotifiedMessages () {
+    return this.latestUnreadMessages.filter((message) => {
+      return parseInt(message.internalDate) > this.lastNotifiedInternalDate
+    })
+  }
+  get lastNotifiedInternalDate () {
+    return this.__data__.unreadMessages.lastNotifiedInternalDate || 0
   }
 }
 
