@@ -2,5 +2,6 @@ module.exports = {
   b64Assets: require('./b64Assets'),
   credentials: require('./credentials'),
   constants: require('./constants'),
-  Models: require('./Models')
+  Models: require('./Models'),
+  dictionaries: require('./dictionaries.js')
 }
